refactor(analysis-types): split supply DTOs
